requirements.txt
# LLM APIs
openai>=1.45.0
anthropic>=0.34.0
google-generativeai>=0.7.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Statistical analysis
scipy>=1.13.0
statsmodels>=0.14.0
pingouin>=0.5.4

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.22.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.32.0

# Testing
pytest>=8.2.0